Reset step output schema when step is removed
