+++ packages/twenty-front/src/modules/workflow/workflow-diagram/components/WorkflowDiagramEffect.tsx
@@ -1,4 +1,5 @@
 import { getSnapshotValue } from '@/ui/utilities/state/utils/getSnapshotValue';
+import { useStepsOutputSchema } from '@/workflow/hooks/useStepsOutputSchema';
 import { flowState } from '@/workflow/states/flowState';
 import { workflowLastCreatedStepIdState } from '@/workflow/states/workflowLastCreatedStepIdState';
 import {
@@ -11,8 +12,9 @@ import { addCreateStepNodes } from '@/workflow/workflow-diagram/utils/addCreateS
 import { getWorkflowVersionDiagram } from '@/workflow/workflow-diagram/utils/getWorkflowVersionDiagram';
 import { markLeafNodes } from '@/workflow/workflow-diagram/utils/markLeafNodes';
 import { mergeWorkflowDiagrams } from '@/workflow/workflow-diagram/utils/mergeWorkflowDiagrams';
+import { TRIGGER_STEP_ID } from '@/workflow/workflow-trigger/constants/TriggerStepId';
 import { useEffect } from 'react';
-import { useRecoilCallback, useSetRecoilState } from 'recoil';
+import { useRecoilCallback, useRecoilState, useSetRecoilState } from 'recoil';
 import { isDefined } from 'twenty-shared';
 
 export const WorkflowDiagramEffect = ({
@@ -21,7 +23,10 @@ export const WorkflowDiagramEffect = ({
   workflowWithCurrentVersion: WorkflowWithCurrentVersion | undefined;
 }) => {
   const setWorkflowDiagram = useSetRecoilState(workflowDiagramState);
-  const setFlow = useSetRecoilState(flowState);
+  const [flow, setFlow] = useRecoilState(flowState);
+  const { resetStepsOutputSchema } = useStepsOutputSchema({
+    instanceIdFromProps: workflowWithCurrentVersion?.currentVersion.id,
+  });
 
   const computeAndMergeNewWorkflowDiagram = useRecoilCallback(
     ({ snapshot, set }) => {
@@ -66,8 +71,8 @@ export const WorkflowDiagramEffect = ({
     [],
   );
 
+  const currentVersion = workflowWithCurrentVersion?.currentVersion;
   useEffect(() => {
-    const currentVersion = workflowWithCurrentVersion?.currentVersion;
     if (!isDefined(currentVersion)) {
       setFlow(undefined);
       setWorkflowDiagram(undefined);
@@ -86,8 +91,34 @@ export const WorkflowDiagramEffect = ({
     computeAndMergeNewWorkflowDiagram,
     setFlow,
     setWorkflowDiagram,
-    workflowWithCurrentVersion?.currentVersion,
+    currentVersion,
   ]);
+
+  useEffect(() => {
+    if (!isDefined(currentVersion)) {
+      return;
+    }
+
+    const deletedSteps = flow?.steps?.filter((flowStep) => {
+      if (!isDefined(currentVersion?.steps)) {
+        return false;
+      }
+
+      return !currentVersion.steps.some(
+        (versionStep) => versionStep.id === flowStep.id,
+      );
+    });
+
+    if (isDefined(deletedSteps)) {
+      deletedSteps.forEach((step) => {
+        resetStepsOutputSchema({ stepId: step.id });
+      });
+    }
+
+    if (!isDefined(currentVersion.trigger) && isDefined(flow?.trigger)) {
+      resetStepsOutputSchema({ stepId: TRIGGER_STEP_ID });
+    }
+  }, [currentVersion, flow?.steps, flow?.trigger, resetStepsOutputSchema]);
 
   return null;
 };
